Milestone3 FINAL TESTING AND FINAL MERGE

+++ frontend/src/components/views/MainNav.js
@@ -9,12 +9,22 @@
 //  (DAB, 03/13/2022, Added in Admin link when an admin is logged in 
 //  light front end only secured)
 //  (DAB, 03/13/2022, Moved DevelopersNav into MainNav)
+//  (DAB, 04/03/2022, Fixed Logout behavior to not return to Login screen 
+//  when clicked)
+//  (DAB, 04/05/2022, Added in sticky nav bar functionality and cleaned up 
+//  code)
+//  (DAB, 4/09/2022, Changed MainNav so that Search Input does not get 
+//  cut off on smaller devices)
+//  (GROUP, 04/14/2022, Search bar now placeholder "Search Restaurants")
+//  (CPD, 4/22/22, Add transition animations when scrolling to hide/show nav)
+//  (CPD, 4/23/22, Fixed position property bug causing nav to vanish when 
+//  loading page form URL)
 
 // Using React library in order to build components 
 // for the app and importing needed components
-import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { Navbar, Button, Nav, Form, Container, FormControl } from 'react-bootstrap';
+import React, { useEffect, useState, useRef } from 'react';
+import { useNavigate, useLocation } from 'react-router-dom';
+import { Navbar, Button, Nav, Form, FormControl } from 'react-bootstrap';
 import { LinkContainer } from 'react-router-bootstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { connect } from 'react-redux';
@@ -25,29 +35,94 @@ import { deleteAllRestaurants } from '../../actions/restaurants';
 import { checkLogin } from '../../helperFunction/CheckLogin';
 import ModalLogoutConfirm from '../modal/LogoutConfirm';
 import DevelopersNav from '../DevelopersNav';
+import checkEnv from "../../helperFunction/checkEnvironment";
 
+// Check if we are on the prod environment
+const isProd = checkEnv();
+
+/**
+ * The MainNav component will allow the user to navigate the 
+ * Restaurant Club Application. It is responsive and will 
+ * change depending on the permission of the user different 
+ * options will be available.
+ * 
+ * @param {*} props 
+ * @returns 
+ */
 function MainNav(props) {
-    // Setting up the basic state needed to run MainNav
-    const [showLogoutConfirm, setShowLogoutConfirm] = useState(false);
-    const [basicActive, setBasicActive] = useState();
-    const [searchInput, setSearchInput] = useState("");
-    const navigate = useNavigate();
-
-    // Get user state from props
+    // Retrieving state data and functions from props
+    const { users } = props;
     const {
-        users,
         deleteAllUsers,
         deleteAllMessages,
         deleteAllReviews,
         deleteAllRestaurants
     } = props;
 
-    // Setting the active nav element
-    const setActive = (value) => {
-        if (value === basicActive) return;
-        console.log("Navigating to " + value)
-        setBasicActive(value);
+
+    // Setting up the basic local state needed to run MainNav
+    const [showLogoutConfirm, setShowLogoutConfirm] = useState(false);
+    const [basicActive, setBasicActive] = useState();
+    const [searchInput, setSearchInput] = useState("");
+    const [navExpanded, setNavExpanded] = useState(false);
+    const [showNav, setShowNav] = useState(true);
+
+    // Refs are used to hold current data within listeners
+    const windowY = useRef(window?.scrollY);
+    const navExpandedRef = useRef(false);
+
+    // Application navigation objects
+    const navigate = useNavigate();
+    const location = useLocation();
+
+    // Theme variables
+    const buttonTheme = "outline-light";
+    const backgroundTheme = "dark";
+    const variantTheme = "dark";
+
+
+    // This use effect will only execute one time when the view is 
+    // initially loaded
+    useEffect(() => {
+        // Adding an event listener to monitor the page scroll
+        window.addEventListener('scroll', hideNavbar);
+
+        return () => {
+            // Unsubscribing to the listener when the view is closed
+            window.removeEventListener('scroll', hideNavbar);
+        }
+    }, []);
+
+
+    // Handler for the LogoutConfirm modal
+    const closeLogoutHandler = () => setShowLogoutConfirm(false);
+
+
+    // The hideNavbar function will either show or hide the navBar 
+    // depending on the windows Y coordinate and weather the user 
+    // is scrolling up or down
+    const hideNavbar = () => {
+        // The navbar will only hide if the navbar is not expanded
+        if (!navExpandedRef.current) {
+            // If the user is scrolling up the navbar will appear 
+            // at the top of the view
+            if (window.scrollY > windowY.current) {
+                // Showing the navbar
+                setShowNav(true)
+                // resetting the last windowY coordinates
+                windowY.current = window.scrollY;
+            }
+            // Else the user is scrolling down so the navbar is 
+            // hidden
+            else {
+                // Hiding the navbar
+                setShowNav(false)
+                // resetting the last windowY coordinates
+                windowY.current = window.scrollY;
+            }
+        }
     }
+
 
     // The searchHandler method will navigate the user 
     // to the search page so a search can be performed
@@ -64,38 +139,85 @@ function MainNav(props) {
 
         // Clearing the active button
         setActive("none")
+
+        // Closing the nav if open
+        if (navExpanded) {
+            setToggle();
+        }
     }
 
-    // Handlers for the LogoutConfirm modal
+
+    // Setting the active nav element
+    const setActive = (value) => {
+        if (value === basicActive) return;
+        setBasicActive(value);
+    }
+
+
+    // The setToggle function works with the navBar and will allow it 
+    // to expand when true or contract when false
+    const setToggle = () => {
+        // Setting the state
+        setNavExpanded(!navExpanded);
+
+        // Setting the ref to be used in the listener
+        navExpandedRef.current = !navExpandedRef.current;
+    }
+
+
+    // Handler for the LogoutConfirm modal
     const showLogoutHandler = () => setShowLogoutConfirm(true);
-    const closeLogoutHandler = () => setShowLogoutConfirm(false);
+
+
+    // Remove everything from state on logout
+    const logoutAccount = () => {
+        // On logout users and user messages are always deleted 
+        // from state
+        deleteAllUsers();
+        deleteAllMessages();
+
+        // If the user is not on the home page, all reviews and 
+        // restaurants are deleted. The home page does this 
+        // already
+        if (location.pathname !== '/') {
+            // Deleting all reviews and restaurants so fresh 
+            // ones can be loaded onto the home page
+            deleteAllReviews();
+            deleteAllRestaurants();
+            // Navigate to home after logout
+            navigate('/');
+        }
+    }
+
 
     // Handles the click to show the modal windows when the logout button is pressed
     const logoutHandler = () => {
+        // Closing the nav if open
+        if (navExpanded) {
+            setToggle();
+        }
+
         // console.log("logout button pressed");
         showLogoutHandler();
     }
 
-    // Remove everything from state on logout
-    const logoutAccount = () => {
-        console.log("logout account");
-        deleteAllUsers();
-        deleteAllMessages();
-        deleteAllReviews();
-        deleteAllRestaurants();
 
-        // Navigate to home after logout
-        navigate("/");
-    }
-
+    // The showAdmin render function will only show the Admin button if the 
+    // user is logged in and had an admin permission
     const showAdmin = () => {
+        // If there is a user in state
         if (users) {
+            // The first user in the array is deconstructed, they are the 
+            // user who's session this is
             const [currentUser] = users;
+
+            // If the user is both logged in and an admin
             if (
                 currentUser.isLoggedIn &&
                 currentUser.auth.permission.permissionName === "admin") {
+                // The Admin button is returned
                 return (
-                    <Nav.Item className="mx-3">
+                    <Nav.Item className="mx-1">
                         <LinkContainer to="/admin">
                             <Nav.Link>
                                 Admin
@@ -107,99 +229,106 @@ function MainNav(props) {
         }
     }
 
+
+    // The showLoginControls render function will render the links 
+    // dependent on if the user is logged in or logged out
     const showLoginControls = () => (
         <>
-            {checkLogin(users) ? (
-                <>
-                    <Nav.Item className="mx-3">
-                        <LinkContainer to="/userDashboard">
-                            <Nav.Link>
-                                Dashboard
-                            </Nav.Link>
-                        </LinkContainer>
-                    </Nav.Item>
-                    {showAdmin()}
-                    <Nav.Item className="mx-3" onClick={logoutHandler}>
-                        <Nav.Link>
-                            Logout
-                        </Nav.Link>
-                    </Nav.Item>
-                </>
-            ) : (
-                <>
-                    <Nav.Item className="mx-3">
-                        <LinkContainer to="/login">
-                            <Nav.Link>
-                                Login
-                            </Nav.Link>
-                        </LinkContainer>
-                    </Nav.Item>
-                </>
-            )}
-
-            <ModalLogoutConfirm
-                show={showLogoutConfirm}
-                logout={logoutAccount}
-                closeHandler={closeLogoutHandler} />
-        </>
-    )
-
-    // Theme variables
-    const buttonTheme = "outline-primary";
-    const backgroundTheme = "light";
-    const variantTheme = "light";
-
-    return (
-        <Container fluid>
-            <Navbar
-                className="mainNav px-2"
-                bg={backgroundTheme}
-                variant={variantTheme}
-                collapseOnSelect
-                expand="md">
-                <Navbar.Brand>
-                    <img
-                        src={window.location.origin + '/logo.gif'}
-                        width="90"
-                        height="90"
-                        className="flex-begin"
-                        alt="Logo"
-                    />
-                </Navbar.Brand>
-                <Navbar.Toggle aria-controls="responsive-navbar-nav" />
-                <Navbar.Collapse id="responsive-navbar-nav" className="justify-content-end">
-                    <Nav fill variant="pills" bg="dark" className="mb-auto pe-3"
-                        activeKey={basicActive}
-                        onSelect={(key) => setActive(key)}>
-                        <DevelopersNav />
-                        <Nav.Item className="mx-3">
-                            <LinkContainer to="/">
+            {checkLogin(users) ?
+                (
+                    <>
+                        <Nav.Item className="mx-1">
+                            <LinkContainer to="/userDashboard">
                                 <Nav.Link>
-                                    Home
+                                    Dashboard
                                 </Nav.Link>
                             </LinkContainer>
                         </Nav.Item>
+                        {showAdmin()}
+                        <Nav.Item className="mx-1" onClick={logoutHandler}>
+                            <Nav.Link>
+                                Logout
+                            </Nav.Link>
+                        </Nav.Item>
+                    </>
+                ) : (
+                    <>
+                        <Nav.Item className="mx-1">
+                            <LinkContainer to="/login">
+                                <Nav.Link>
+                                    Login
+                                </Nav.Link>
+                            </LinkContainer>
+                        </Nav.Item>
+                    </>
+                )
+            }
+        </>
+    )
 
-                        {showLoginControls()}
 
-                    </Nav>
-                    <Form onSubmit={searchHandler} className="d-flex">
-                        <FormControl
-                            type="search"
-                            name="searchInput"
-                            value={searchInput}
-                            onInput={e => setSearchInput(e.target.value)}
-                            placeholder="Search"
-                            className="me-2"
-                            aria-label="Search"
-                        />
-                        <Button type="submit" variant={buttonTheme}>
-                            Search
-                        </Button>
-                    </Form>
-                </Navbar.Collapse>
-            </Navbar>
-        </Container>
+    return (
+        <Navbar
+            id="mainNav"
+            onToggle={setToggle}
+            className="px-2 rounded-bottom"
+
+            style={showNav ? 
+                { position: "sticky", top: "-120px", zIndex: "1", transition: 'top 0.6s' } :
+                { position: "sticky", top: "0", bottom: "0", zIndex: "1", transition: 'top 0.6s' }
+            }
+            bg={backgroundTheme}
+            variant={variantTheme}
+            collapseOnSelect
+            expanded={navExpanded}
+            expand="md">
+            <Navbar.Brand>
+                <img
+                    src={window.location.origin + '/logo.gif'}
+                    width="124px"
+                    height="90px"
+                    className="flex-begin rounded"
+                    alt="Main Logo"
+                />
+            </Navbar.Brand>
+            <Navbar.Toggle aria-controls="responsive-navbar-nav" />
+            <Navbar.Collapse id="responsive-navbar-nav" className="justify-content-end flex-shrink-1" >
+                <Nav fill className="mb-auto"
+                    activeKey={location.pathname.substring(0, location.pathname.lastIndexOf('/'))}
+                    onSelect={(key) => setActive(key)}>
+
+                    {/* Show dev nav if we are not on prod env */}
+                    {(isProd === false) && <DevelopersNav />}
+
+                    <Nav.Item className="mx-1">
+                        <LinkContainer to="/">
+                            <Nav.Link>
+                                Home
+                            </Nav.Link>
+                        </LinkContainer>
+                    </Nav.Item>
+                    {showLoginControls()}
+                </Nav>
+                <Form onSubmit={searchHandler} className="d-flex" style={{ minWidth: "10rem" }}>
+                    <FormControl
+                        type="search"
+                        name="searchInput"
+                        value={searchInput}
+                        onInput={e => setSearchInput(e.target.value)}
+                        placeholder="Search Restaurants"
+                        className="me-2"
+                        aria-label="Search Restaurants"
+                    />
+                    <Button type="submit" variant={buttonTheme} >
+                        Search
+                    </Button>
+                </Form>
+                <ModalLogoutConfirm
+                    show={showLogoutConfirm}
+                    logout={logoutAccount}
+                    closeHandler={closeLogoutHandler} />
+            </Navbar.Collapse>
+        </Navbar>
     )
 }
